Added optional ObjectConstraintRegistry to ObjectConstraintsBuilder, changed ClosureObjectConstraintRegistrant to be ObjectConstraintsBuilderRegistrant so that we can use builders to register constraints (better DX)
