+++ src/Validation/src/Builders/ObjectConstraintsBuilderRegistrant.php
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * Aphiria
+ *
+ * @link      https://www.aphiria.com
+ * @copyright Copyright (C) 2019 David Young
+ * @license   https://github.com/aphiria/aphiria/blob/master/LICENSE.md
+ */
+
+declare(strict_types=1);
+
+namespace Aphiria\Validation\Builders;
+
+use Aphiria\Validation\Constraints\IObjectConstraintRegistrant;
+use Aphiria\Validation\Constraints\ObjectConstraintRegistry;
+use Closure;
+
+/**
+ * Defines the constraint registrant that populates the object constraints using builders
+ */
+final class ObjectConstraintsBuilderRegistrant implements IObjectConstraintRegistrant
+{
+    /** @var Closure[] The list of closures to execute (must take in an ObjectConstraintsBuilder parameter) */
+    private array $closures;
+
+    /**
+     * @param Closure[] $closures The list of closures to execute (must take in an ObjectConstraintsBuilder parameter)
+     */
+    public function __construct(array $closures)
+    {
+        $this->closures = $closures;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function registerConstraints(ObjectConstraintRegistry $objectConstraints): void
+    {
+        $objectConstraintsBuilder = new ObjectConstraintsBuilder($objectConstraints);
+
+        foreach ($this->closures as $closure) {
+            $closure($objectConstraintsBuilder);
+        }
+
+        $objectConstraintsBuilder->build();
+    }
+}
